+++ src/routes/AdminRoute.js
@@ -1,0 +1,25 @@
+import React from "react";
+import { Route, Redirect } from "react-router-dom";
+
+import { isAuthenticated, isAdmin } from "../services/auth";
+
+const AdminRoute = ({ component: Component, ...rest }) => (
+  <Route
+    {...rest}
+    render={props =>
+      isAuthenticated() ? (
+        isAdmin() ? (
+          <Component {...props} />
+        ) : (
+          <Redirect to={{ pathname: "/", state: { from: props.location } }} />
+        )
+      ) : (
+        <Redirect
+          to={{ pathname: "/login", state: { from: props.location } }}
+        />
+      )
+    }
+  />
+);
+
+export default AdminRoute;
