+++ src/routes/index.js
@@ -1,49 +1,17 @@
 import React from "react";
-import { Router, Route, Switch, Redirect } from "react-router-dom";
+import { Router, Route, Switch } from "react-router-dom";
+import history from "../history";
 
 import Login from "../components/auth/Login";
 import AdminList from "../components/admins/AdminList";
 import AdminCreate from "../components/admins/AdminCreate";
 import AdminEdit from "../components/admins/AdminEdit";
 import AdminDelete from "../components/admins/AdminDelete";
-import history from "../history";
 import Header from "../components/Header";
 import FlashMessage from "../components/FlashMessage";
-import { isAuthenticated, isAdmin } from "../apis/auth";
 
-const PrivateRoute = ({ component: Component, ...rest }) => (
-  <Route
-    {...rest}
-    render={props =>
-      isAuthenticated() ? (
-        <Component {...props} />
-      ) : (
-        <Redirect
-          to={{ pathname: "/login", state: { from: props.location } }}
-        />
-      )
-    }
-  />
-);
-
-const AdminRoute = ({ component: Component, ...rest }) => (
-  <Route
-    {...rest}
-    render={props =>
-      isAuthenticated() ? (
-        isAdmin() ? (
-          <Component {...props} />
-        ) : (
-          <Redirect to={{ pathname: "/", state: { from: props.location } }} />
-        )
-      ) : (
-        <Redirect
-          to={{ pathname: "/login", state: { from: props.location } }}
-        />
-      )
-    }
-  />
-);
+import PrivateRoute from "./PrivateRoute";
+import AdminRoute from "./AdminRoute";
 
 const Routes = () => (
   <div className="ui container">
@@ -54,12 +22,12 @@ const Routes = () => (
         <Switch>
           <PrivateRoute exact path="/" component={() => <h1>Dashboard</h1>} />
 
-          <AdminRoute path="/admins" exact component={AdminList} />
-          <AdminRoute path="/admins/new" exact component={AdminCreate} />
-          <AdminRoute path="/admins/:id/edit" exact component={AdminEdit} />
-          <AdminRoute path="/admins/:id/delete" exact component={AdminDelete} />
+          <AdminRoute exact path="/admins" component={AdminList} />
+          <AdminRoute exact path="/admins/new" component={AdminCreate} />
+          <AdminRoute exact path="/admins/:id/edit" component={AdminEdit} />
+          <AdminRoute exact path="/admins/:id/delete" component={AdminDelete} />
 
-          <Route path="/login" exact component={Login} />
+          <Route exact path="/login" component={Login} />
           <Route path="*" component={() => <h1>Page not found</h1>} />
         </Switch>
       </div>
